feat: introduce emotion
