+++ packages/blocky-react/tsconfig.json
@@ -18,6 +18,7 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "react-jsx",
+    "jsxImportSource": "@emotion/react",
     "declaration": true,
     "declarationDir": "./dist",
     "inlineSourceMap": true,
